Update file names and separate logic out
