+++ sample-apps/Dockerfile
@@ -4,6 +4,6 @@ WORKDIR /usr/src/app
 
 COPY . .
 
-RUN cd sample-apps && npm install && cd node_modules/AWSXRayIdGenerator && npm install && npm run compile && cd ../AWSXRayPropagator && npm install && npm run compile
+RUN npm install
 
-CMD ["node", "sample-apps/server.js" ]
+CMD ["node", "server.js" ]
